add base16 colorscheme.
I sure hope this works on all my other terminals, fuck. Also the current
configuration only works if I start a tmux session, though after I exit
the session it continues to work.

You can make the configuration work before a tmux session is entered, by
commenting out "let base16colorspace=256" in vimrc. But in that case it
ceases to work after the tmux session is started.

tl;dr: fuck, just fucking use tmux and hope for the best. This
colorscheme is really awesome but the setup is a pain.